requirements.txt
pandas == 1.5.2
numpy == 1.21.5
matplotlib == 3.5.2
seaborn == 0.11.2
scipy == 1.9.1
apache-airflow == 2.5.0
dash == 2.7.1
dash-core-components == 2.0.0
SQLAlchemy == 1.4.45
